# docs/swagger.yaml
basePath: /
definitions:
  link.CreateLinkRequest:
    properties:
      url:
        type: string
    type: object
  link.LinkResponse:
    properties:
      created_at:
        type: string
      link_body:
        type: string
      link_code:
        type: string
    type: object
  statistics.StatisticsResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      redirects_count:
        type: integer
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for URL shortening service
  title: ShortenIt API
  version: "1.0"
paths:
  /shorten:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/link.LinkResponse'
            type: array
        "500":
          description: Internal Server Error
      summary: Returns the list of all accessable URLs.
    post:
      consumes:
      - application/json
      parameters:
      - description: URL
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/link.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.LinkResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/link.LinkResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Creates a new short code based on a provided link. If such a code already
        exists, returns an existing record.
  /shorten/{code}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.LinkResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Deletes a URL by its short code.
    get:
      responses:
        "320":
          description: ""
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Processes redirect to a URL by its short code.
  /shorten/{code}/statistics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/statistics.StatisticsResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Returns statistics related to a URL by its short code.
swagger: "2.0"